+++ tests/integration/storage/SessionTest.php
@@ -10,6 +10,7 @@
 namespace lithium\tests\integration\storage;
 
 use lithium\storage\Session;
+use lithium\storage\session\strategy\Encrypt;
 
 class SessionTest extends \lithium\test\Integration {
 
@@ -121,95 +122,6 @@ class SessionTest extends \lithium\test\Integration {
 		$this->assertEqual($value2, Session::read($key2, $config));
 		$expected = ['nested' => ['key' => $value1]];
 		$this->assertEqual($expected, Session::read('really.deep', $config));
-	}
-
-	public function testHmacStrategyWithPhpAdapter() {
-		$this->skipIf(PHP_SAPI === 'cli', 'No PHP session support in cli SAPI.');
-
-		$config = ['name' => 'hmacInt'];
-
-		Session::config([
-			$config['name'] => [
-				'adapter' => 'Php',
-				'strategies' => [
-					'Hmac' => [
-						'secret' => 's3cr3t'
-					]
-				]
-			]
-		]);
-
-		Session::clear($config);
-
-		$key = 'test';
-		$value = 'value';
-
-		$this->assertTrue(Session::write($key, $value, $config));
-		$this->assertEqual($value, Session::read($key, $config));
-		$this->assertTrue(Session::delete($key, $config));
-		$this->assertNull(Session::read($key, $config));
-
-		Session::clear($config);
-
-		$this->assertTrue(Session::write('foo', 'bar', $config));
-		$this->assertEqual('bar', Session::read('foo', $config));
-		$this->assertTrue(Session::write('foo', 'bar1', $config));
-		$this->assertEqual('bar1', Session::read('foo', $config));
-
-		Session::clear($config);
-
-		$this->assertTrue(Session::write($key, $value, $config));
-		$this->assertEqual($value, Session::read($key, $config));
-
-		$cache = $_SESSION;
-		$_SESSION['injectedkey'] = 'hax0r';
-		$expected = '/Possible data tampering: HMAC signature does not match data./';
-		$this->asssertException($expected, function() use ($key, $config) {
-			Session::read($key, $config);
-		});
-		$_SESSION = $cache;
-
-		Session::reset();
-	}
-
-	public function testEncryptStrategyWithPhpAdapter() {
-		$this->skipIf(PHP_SAPI === 'cli', 'No PHP session support in cli SAPI.');
-		$this->skipIf(!extension_loaded('mcrypt'), 'The `mcrypt` extension is not loaded.');
-
-		$config = ['name' => 'encryptInt'];
-
-		Session::config([
-			$config['name'] => [
-				'adapter' => 'Php',
-				'strategies' => [
-					'Encrypt' => [
-						'secret' => 's3cr3t'
-					]
-				]
-			]
-		]);
-
-		Session::clear($config);
-
-		$key = 'test';
-		$value = 'value';
-
-		$this->assertTrue(Session::write($key, $value, $config));
-		$this->assertEqual($value, Session::read($key, $config));
-		$this->assertTrue(Session::delete($key, $config));
-		$this->assertNull(Session::read($key, $config));
-
-		Session::clear($config);
-
-		$this->assertTrue(Session::write('foo', 'bar', $config));
-		$this->assertEqual('bar', Session::read('foo', $config));
-		$this->assertTrue(Session::write('foo', 'bar1', $config));
-		$this->assertEqual('bar1', Session::read('foo', $config));
-
-		Session::clear($config);
-
-		$this->assertTrue(Session::write($key, $value, $config));
-		$this->assertEqual($value, Session::read($key, $config));
 	}
 }